Фантастика, а ти міг би перевірити вивід іконок у картках головної у сек
